refactor(ComposeMessageView): Move attachments to their own view + Use separate object for attachment management
